radicados-bot/requirements.txt
pymongo==4.6.1
PyPDF2==3.0.1
python-telegram-bot==20.7
python-dotenv==1.0.0
pycryptodome==3.23.0
